Time: 3 ms (71.20%) | Memory: 10.7 MB (88.53%) - LeetSync

+++ 14-longest-common-prefix/longest-common-prefix.cpp
@@ -5,7 +5,7 @@ public:
         if (strs.empty()) return "";
         string ans;
         while (1) {
-            if (++j >= (int) strs[0].size()) break;
+            if (++j >= strs[0].size()) break;
             char c = strs[0][j];
             for (int i = 1; i < sz; ++i) {
                 if (strs[i].size() <= j || strs[i][j] != c) return ans;
